Create CODE
4-bit PIPO (Parallel-In, Parallel-Out) shift register:-
This code defines a Verilog module named pipo. 
It takes a 4-bit input pin, clock signal clk, and reset signal rst . The output is also a 4-bit signal pout.

Within the always block sensitive to the positive edge of the clock (posedge clk), the input pin is loaded into the output pout. When the reset signal rst is asserted, the output pout is set to all zeros.